6g-monitoring/consumer/requirements.txt
prometheus-client==0.20.0
confluent-kafka==2.3.0
pyyaml==6.0.1
